+++ NEWS.md
@@ -1,3 +1,11 @@
+# uklakes (development version)
+
+* removing biology table check as seems redundant
+
+* adding extra test for `search_buffer`
+
+* adding check for internet connectivity (`curl` dependency)
+
 # uklakes 0.2.0
 
 * Adding in `search_buffer` function to allow search for lakes within given distance from a point.